+++ __init__.py
@@ -5,7 +5,7 @@ from .nodes.text_nodes import *
 from .nodes.key_control_nodes import *
 from .nodes.osc_control_nodes import *
 
-__version__ = "1.0.19"
+__version__ = "1.0.20"
 
 NODE_CLASS_MAPPINGS = {
     "VrchAnyOSCControlNode": VrchAnyOSCControlNode,
@@ -20,9 +20,11 @@ NODE_CLASS_MAPPINGS = {
     "VrchFloatKeyControlNode": VrchFloatKeyControlNode,
     "VrchFloatOSCControlNode": VrchFloatOSCControlNode,
     "VrchJsonUrlLoaderNode": VrchJsonUrlLoaderNode,
+    "VrchImageChannelLoaderNode": VrchImageChannelLoaderNode,
     "VrchImageFlipBookWebViewerNode": VrchImageFlipBookWebViewerNode,
     "VrchImageSaverNode": VrchImageSaverNode,
     "VrchImageSwitchOSCControlNode": VrchImageSwitchOSCControlNode,
+    "VrchImagePreviewBackgroundNode": VrchImagePreviewBackgroundNode,
     "VrchImageWebViewerNode": VrchImageWebViewerNode,
     "VrchIntKeyControlNode": VrchIntKeyControlNode,
     "VrchIntOSCControlNode": VrchIntOSCControlNode,
@@ -55,7 +57,9 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "VrchImageFlipBookWebViewerNode": "IMAGE Flipbook Web Viewer @vrch.ai",
     "VrchImageSaverNode": "IMAGE Saver @ vrch.ai",
     "VrchImageSwitchOSCControlNode": "IMAGE Switch OSC Control @ vrch.ai",
+    "VrchImagePreviewBackgroundNode": "IMAGE Preview in Background @ vrch.ai",
     "VrchImageWebViewerNode": "IMAGE Web Viewer @ vrch.ai",
+    "VrchImageChannelLoaderNode": "IMAGE Web Viewer Channel Loader @ vrch.ai",
     "VrchIntKeyControlNode": "INT Key Control @ vrch.ai",
     "VrchIntOSCControlNode": "INT OSC Control @ vrch.ai",
     "VrchInstantQueueKeyControlNode": "Instant Queue Key Control @ vrch.ai",